Refactor asset quality helpers into modular utilities
